[ feat ] - messages

+++ controllers/message.controller.js
@@ -3,29 +3,32 @@ const websocket = require( '../utils/websocket/websocket' );
 const apiResponse = require( '../utils/helpers/api.response' );
 
 exports.index = async ( req, res ) => {
-    try {
-        // Extract pagination parameters from query string
-        let { page = 1, perPage = 10 } = req.query;
+    try
+    {
+        let { page = 1, perPage = 10, ...filters } = req.query;
 
-        // Convert to integers and ensure valid numbers
-        page = Math.max( parseInt( page ), 1 ); // Ensure page is at least 1
-        perPage = Math.max( parseInt( perPage ), 1 ); // Ensure limit is at least 1
-
-        // Calculate how many documents to skip
+        page = Math.max( parseInt( page ), 1 );
+        perPage = Math.max( parseInt( perPage ), 1 );
         const skip = ( page - 1 ) * perPage;
 
-        // Fetch paginated messages
-        const messages = await Message.find()
+        // Construct filter object dynamically
+        let filter = {};
+        for ( const key in filters )
+        {
+            if ( filters[ key ] )
+            {
+                filter[ key ] = filters[ key ];
+            }
+        }
+
+        const messages = await Message.find( filter )
             .skip( skip )
             .limit( perPage );
 
-        // Count total documents (for pagination info)
-        const totalItems = await Message.countDocuments();
-
-        // Calculate total pages
+        const totalItems = await Message.countDocuments( filter );
         const totalPages = Math.ceil( totalItems / perPage );
 
-        apiResponse( res, 'success', 'Message has indexed successfully.', {
+        apiResponse( res, 'success', 'Message has been indexed successfully.', {
             items: messages,
             page: page,
             perPage: perPage,
@@ -55,21 +58,28 @@ exports.show = async ( req, res ) => {
 
 exports.update = async ( req, res ) => {
     try {
-        const message = await Message.findByIdAndUpdate( req.params.id, req.body, {
-            new: true,
-            runValidators: true // Enables schema validation on update
-        } );
-
+        const message = await Message.findById( req.params.id );
         if ( !message )
         {
             throw new Error( 'Message not found.' );
         }
 
+        // Check if the sender is the same as the requesting user
+        if ( message.sender_id !== req.user.id )
+        {
+            throw new Error( 'Unauthorized.' );
+        }
+
+        const updatedMessage = await Message.findByIdAndUpdate( req.params.id, req.body, {
+            new: true,
+            runValidators: true
+        } );
+
         // Notify users via WebSocket
         const io = websocket.getIO();
-        io.to( message.room_id ).emit( 'edit-message', message );
+        io.to( message.room_id ).emit( 'edit_message', updatedMessage );
 
-        apiResponse( res, 'success', 'Message has updated successfully.', message );
+        apiResponse( res, 'success', 'Message has updated successfully.', updatedMessage );
     }
     catch ( error ) {
         apiResponse( res, 'failed', 'Bad request.', error.toString(), 400 );
@@ -78,17 +88,25 @@ exports.update = async ( req, res ) => {
 
 exports.delete = async ( req, res ) => {
     try {
-        const message = await Message.findByIdAndDelete( req.params.id );
+        const message = await Message.findById( req.params.id );
         if ( !message )
         {
             throw new Error( 'Message not found.' );
         }
 
+        // Check if the sender is the same as the requesting user
+        if ( message.sender_id !== req.user.id )
+        {
+            throw new Error( 'Unauthorized.' );
+        }
+
+        const deletedMessage = await Message.findByIdAndDelete( req.params.id );
+
         // Notify users via WebSocket
         const io = websocket.getIO();
-        io.to( message.room_id ).emit( 'delete-message', message.id );
+        io.to( message.room_id ).emit( 'delete_message', deletedMessage.id );
 
-        apiResponse( res, 'success', 'Message has deleted successfully.', message );
+        apiResponse( res, 'success', 'Message has deleted successfully.', deletedMessage );
     }
     catch ( error ) {
         apiResponse( res, 'failed', 'Bad request.', error.toString(), 400 );
